+++ src/components/petcard/index.ts
@@ -1,2 +1,4 @@
+export * from './generateItem';
+export * from './generateview';
 export * from './progressbar';
 export * from './tab';
